fix(project) reusable project form
