+++ src/pages/projects/EditProjectForm.tsx
@@ -1,20 +1,12 @@
 import React, { FC, useEffect, useState } from "react";
-import { Button, Col, Form, Row } from "@canonical/react-components";
-import NotificationRow from "components/NotificationRow";
+import { Button, Col, Row } from "@canonical/react-components";
 import { updateProject } from "api/projects";
 import { useQueryClient } from "@tanstack/react-query";
 import { queryKeys } from "util/queryKeys";
 import { useNotify } from "context/notify";
 import DeleteProjectBtn from "pages/projects/actions/DeleteProjectBtn";
-import ProjectFormMenu, {
-  CLUSTERS,
-  DEVICE_USAGE,
-  INSTANCES,
-  NETWORKS,
-  PROJECT_DETAILS,
-  RESOURCE_LIMITS,
-} from "pages/projects/forms/ProjectFormMenu";
-import ProjectDetailsForm, {
+import { PROJECT_DETAILS } from "pages/projects/forms/ProjectFormMenu";
+import {
   projectDetailPayload,
   projectDetailRestrictionPayload,
 } from "pages/projects/forms/ProjectDetailsForm";
@@ -25,24 +17,15 @@ import * as Yup from "yup";
 import { LxdProject } from "types/project";
 import { updateMaxHeight } from "util/updateMaxHeight";
 import useEventListener from "@use-it/event-listener";
-import ResourceLimitsForm, {
-  resourceLimitsPayload,
-} from "pages/projects/forms/ResourceLimitsForm";
-import ClusterRestrictionForm, {
-  clusterRestrictionPayload,
-} from "pages/projects/forms/ClusterRestrictionForm";
-import InstanceRestrictionForm, {
-  instanceRestrictionPayload,
-} from "pages/projects/forms/InstanceRestrictionForm";
-import DeviceUsageRestrictionForm, {
-  deviceUsageRestrictionPayload,
-} from "pages/projects/forms/DeviceUsageRestrictionForm";
-import NetworkRestrictionForm, {
-  networkRestrictionPayload,
-} from "pages/projects/forms/NetworkRestrictionForm";
+import { resourceLimitsPayload } from "pages/projects/forms/ResourceLimitsForm";
+import { clusterRestrictionPayload } from "pages/projects/forms/ClusterRestrictionForm";
+import { instanceRestrictionPayload } from "pages/projects/forms/InstanceRestrictionForm";
+import { deviceUsageRestrictionPayload } from "pages/projects/forms/DeviceUsageRestrictionForm";
+import { networkRestrictionPayload } from "pages/projects/forms/NetworkRestrictionForm";
 import classnames from "classnames";
 import { getProjectEditValues } from "util/projectEdit";
 import { FormikProps } from "formik/dist/types";
+import ProjectForm from "pages/projects/forms/ProjectForm";
 
 interface Props {
   project: LxdProject;
@@ -52,7 +35,6 @@ const EditProjectForm: FC<Props> = ({ project }) => {
   const notify = useNotify();
   const queryClient = useQueryClient();
   const [section, setSection] = useState(PROJECT_DETAILS);
-  const [isRestrictionsOpen, setRestrictionsOpen] = useState(false);
 
   const updateFormHeight = () => {
     updateMaxHeight("form-contents", "p-bottom-controls");
@@ -106,14 +88,6 @@ const EditProjectForm: FC<Props> = ({ project }) => {
     },
   });
 
-  const updateSection = (newItem: string) => {
-    setSection(newItem);
-  };
-
-  const toggleMenu = () => {
-    setRestrictionsOpen((old) => !old);
-  };
-
   return (
     <main className="l-main">
       <div className="p-panel">
@@ -121,40 +95,11 @@ const EditProjectForm: FC<Props> = ({ project }) => {
           <h4 className="p-panel__title">Project configuration</h4>
         </div>
         <div className="p-panel__content edit-project">
-          <Form onSubmit={formik.handleSubmit} stacked className="form">
-            <ProjectFormMenu
-              active={section}
-              setActive={updateSection}
-              isRestrictionsOpen={
-                isRestrictionsOpen && formik.values.restricted
-              }
-              isRestrictionsDisabled={!formik.values.restricted}
-              toggleRestrictionsOpen={toggleMenu}
-            />
-            <Row className="form-contents" key={section}>
-              <Col size={12}>
-                <NotificationRow />
-                {section === PROJECT_DETAILS && (
-                  <ProjectDetailsForm formik={formik} />
-                )}
-                {section === RESOURCE_LIMITS && (
-                  <ResourceLimitsForm formik={formik} />
-                )}
-                {section === CLUSTERS && (
-                  <ClusterRestrictionForm formik={formik} />
-                )}
-                {section === INSTANCES && (
-                  <InstanceRestrictionForm formik={formik} />
-                )}
-                {section === DEVICE_USAGE && (
-                  <DeviceUsageRestrictionForm formik={formik} />
-                )}
-                {section === NETWORKS && (
-                  <NetworkRestrictionForm formik={formik} />
-                )}
-              </Col>
-            </Row>
-          </Form>
+          <ProjectForm
+            formik={formik}
+            section={section}
+            updateSection={setSection}
+          />
           <div className="p-bottom-controls">
             <hr />
             <Row>
